adding signup

+++ src/pages/register.vue
@@ -1,12 +1,11 @@
 <script setup lang="ts">
-import { VForm } from 'vuetify/components/VForm'
-import type { RegisterResponse } from '@/@fake-db/types'
 import authV2RegisterIllustrationBorderedDark from '@images/pages/auth-v2-register-illustration-bordered-dark.png'
 import authV2RegisterIllustrationBorderedLight from '@images/pages/auth-v2-register-illustration-bordered-light.png'
 import authV2RegisterIllustrationDark from '@images/pages/auth-v2-register-illustration-dark.png'
 import authV2RegisterIllustrationLight from '@images/pages/auth-v2-register-illustration-light.png'
 import authV2MaskDark from '@images/pages/misc-mask-dark.png'
 import authV2MaskLight from '@images/pages/misc-mask-light.png'
+import { VForm } from 'vuetify/components/VForm'
 
 import { useAppAbility } from '@/plugins/casl/useAppAbility'
 import AuthProvider from '@/views/pages/authentication/AuthProvider.vue'
@@ -17,9 +16,13 @@ import { themeConfig } from '@themeConfig'
 import { alphaDashValidator, emailValidator, requiredValidator } from '@validators'
 
 const refVForm = ref<VForm>()
-const username = ref('johnDoe')
-const email = ref('john@example.com')
-const password = ref('john@VUEXY#123')
+const username = ref('test123')
+const email = ref('test123@gmail.com')
+const password = ref('soccer')
+const confirmPassword = ref('soccer')
+const workspaceName = ref('test123')
+const isLoading = ref(false)
+
 const privacyPolicies = ref(true)
 
 // Router
@@ -33,33 +36,41 @@ const ability = useAppAbility()
 const errors = ref<Record<string, string | undefined>>({
   email: undefined,
   password: undefined,
+  confirm_password: undefined,
 })
 
 const register = () => {
-  axios.post<RegisterResponse>('/auth/register', {
-    username: username.value,
+  isLoading.value = true
+  axios.post('/v1/signup', {
     email: email.value,
     password: password.value,
+    confirm_password: confirmPassword.value,
+    username: username.value,
+    workspace_name: workspaceName.value,
   })
-    .then(r => {
-      const { accessToken, userData, userAbilities } = r.data
+    .then(res => {
+      const { session: accessToken, user: userData, user_workspace: userWorkspace } = res.data
 
+      const userAbilities = [{ action:"manage",subject:"all" }]
       localStorage.setItem('userAbilities', JSON.stringify(userAbilities))
       ability.update(userAbilities)
 
       localStorage.setItem('userData', JSON.stringify(userData))
       localStorage.setItem('accessToken', JSON.stringify(accessToken))
+      localStorage.setItem('userWorkspace', JSON.stringify(userWorkspace))
+
 
       // Redirect to `to` query if exist or redirect to index route
       router.replace(route.query.to ? String(route.query.to) : '/')
 
-      return null
     })
     .catch(e => {
-      const { errors: formErrors } = e.response.data
+      const error = e?.response?.data || {}
 
-      errors.value = formErrors
+      // errors.value = formErrors
       console.error(e.response.data)
+    }).finally(() => {
+      isLoading.value = false
     })
 }
 
@@ -124,10 +135,10 @@ const onSubmit = () => {
             class="mb-6"
           />
           <h5 class="text-h5 mb-1">
-            Adventure starts here 🚀
+            Future starts here 🚀
           </h5>
           <p class="mb-0">
-            Make your app management easy and fun!
+            Streamline and automate your communications
           </p>
         </VCardText>
 
@@ -138,7 +149,7 @@ const onSubmit = () => {
           >
             <VRow>
               <!-- Username -->
-              <VCol cols="12">
+              <VCol cols="12" class="py-1">
                 <AppTextField
                   v-model="username"
                   autofocus
@@ -148,7 +159,7 @@ const onSubmit = () => {
               </VCol>
 
               <!-- email -->
-              <VCol cols="12">
+              <VCol cols="12" class="py-1">
                 <AppTextField
                   v-model="email"
                   :rules="[requiredValidator, emailValidator]"
@@ -157,8 +168,15 @@ const onSubmit = () => {
                 />
               </VCol>
 
-              <!-- password -->
-              <VCol cols="12">
+              <VCol cols="12" class="py-1">
+                <AppTextField
+                  v-model="workspaceName"
+                  :rules="[requiredValidator]"
+                  label="Workspace name"
+                />
+              </VCol>
+
+              <VCol cols="12" class="py-1">
                 <AppTextField
                   v-model="password"
                   :rules="[requiredValidator]"
@@ -167,6 +185,21 @@ const onSubmit = () => {
                   :append-inner-icon="isPasswordVisible ? 'tabler-eye-off' : 'tabler-eye'"
                   @click:append-inner="isPasswordVisible = !isPasswordVisible"
                 />
+              </VCol>
+
+              <VCol cols="12" class="py-1">
+                <AppTextField
+                  v-model="confirmPassword"
+                  :rules="[requiredValidator]"
+                  label="Confirm Password"
+                  :type="isPasswordVisible ? 'text' : 'password'"
+                  :append-inner-icon="isPasswordVisible ? 'tabler-eye-off' : 'tabler-eye'"
+                  @click:append-inner="isPasswordVisible = !isPasswordVisible"
+                />
+              </VCol>
+
+              <!-- password -->
+              <VCol cols="12">
 
                 <div class="d-flex align-center mt-2 mb-4">
                   <VCheckbox
@@ -190,6 +223,7 @@ const onSubmit = () => {
                 <VBtn
                   block
                   type="submit"
+                  :loading="isLoading"
                 >
                   Sign up
                 </VBtn>
